+++ compiler/compiler_typing/src/lib.rs
@@ -14,6 +14,7 @@ pub mod enums;
 pub mod references;
 pub mod utils;
 pub mod storage;
+pub mod transmutation;
 
 /// A function contained within a type.
 pub type TypedFunction = (Vec<(u64, TypeReference)>, Option<TypeReference>);
